Move CTFd API -> DB calls to own file
