+++ ctfdb/src/ctfd/mod.rs
@@ -1,2 +1,3 @@
 pub mod api;
 pub mod structs;
+pub mod db;
